structure added to firebase functions
